+++ directanswercards/card_component.js
@@ -5,17 +5,13 @@ BaseDirectAnswerCard = typeof (BaseDirectAnswerCard) !== 'undefined' ?
 BaseDirectAnswerCard["{{componentName}}"] = class extends ANSWERS.Component {
   constructor(config = {}, systemConfig = {}) {
     super(config, systemConfig);
-    let data = config.data || {};
+    const data = config.data || {};
     this.type = data.type || '';
     this.answer = data.answer || {};
     this.snippet = this.answer.snippet || {};
     this.relatedItem = data.relatedItem || {};
     this.associatedEntityId = data.relatedItem && data.relatedItem.data && data.relatedItem.data.id;
     this.verticalConfigId = data.relatedItem && data.relatedItem.verticalConfigId;
-    this.globalOptions = this.analyticsReporter && this.analyticsReporter._globalOptions;
-    this.experienceKey = this.globalOptions && this.globalOptions.experienceKey;
-    this.experienceVersion = this.globalOptions && this.globalOptions.experienceVersion;
-    this.queryId = this.globalOptions && this.globalOptions.queryId;
   }
 
   /**
@@ -44,15 +40,22 @@ BaseDirectAnswerCard["{{componentName}}"] = class extends ANSWERS.Component {
   }
 
   onMount() {
-    let feedbackFormSelector = '.js-HitchhikerCard-feedbackForm';
+    this.addFeedbackListeners();
+
+    const rtfElement = this._container.querySelector('.js-yxt-rtfValue');
+    rtfElement && rtfElement.addEventListener('click', e => this._handleRtfClickAnalytics(e));
+  }
+
+  addFeedbackListeners() {
+    const feedbackFormSelector = '.js-HitchhikerCard-feedbackForm';
     let feedbackFormEl = this._container.querySelector(feedbackFormSelector);
     if (feedbackFormEl) {
       // For WCAG compliance, the feedback should be a submittable form
       feedbackFormEl.addEventListener('submit', (e) => {
-        let formTargetEl = e.target;
-        let checkedValue = formTargetEl.querySelector('input:checked').value === 'true';
+        const formTargetEl = e.target;
+        const isGood = formTargetEl.querySelector('input:checked').value === 'true';
 
-        this.reportQuality(checkedValue);
+        this.reportQuality(isGood);
         this.updateState({
           feedbackSubmitted: true
         });
@@ -66,42 +69,11 @@ BaseDirectAnswerCard["{{componentName}}"] = class extends ANSWERS.Component {
             if (input) {
               input.checked = true;
             }
-            this._triggerCustomEvent(feedbackFormSelector, 'submit');
+            HitchhikerJS.DOM.triggerCustomEvent(this._container, feedbackFormSelector, 'submit');
           });
         });
       }
     }
-
-    const rtfElement = this._container.querySelector('.js-yxt-rtfValue');
-    rtfElement && rtfElement.addEventListener('click', e => this._handleRtfClickAnalytics(e));
-  }
-
-  /**
-   * Triggers the event passed in dispatched from the given selector
-   * @param {string} selector selector to dispatch event from
-   * @param {string} event event to fire
-   * @param {Object} settings additional settings
-   */
-  _triggerCustomEvent(selector, event, settings) {
-    let e = this._customEvent(event, settings);
-    this._container.querySelector(selector).dispatchEvent(e);
-  }
-
-  /**
-   * _customEvent is an event constructor polyfill
-   * @param {string} event event to fire
-   * @param {Object} settings additional settings
-   */
-  _customEvent(event, settings) {
-    const _settings = {
-      bubbles: true,
-      cancelable: true,
-      detail: null,
-      ...settings
-    };
-    const evt = document.createEvent('CustomEvent');
-    evt.initCustomEvent(event, _settings.bubbles, _settings.cancelable, _settings.detail);
-    return evt;
   }
 
   /**
@@ -123,9 +95,6 @@ BaseDirectAnswerCard["{{componentName}}"] = class extends ANSWERS.Component {
     const event = new ANSWERS.AnalyticsEvent(eventType)
       .addOptions({
         directAnswer: true,
-        experienceKey: this.experienceKey,
-        experienceVersion: this.experienceVersion,
-        queryId: this.queryId,
         verticalKey: this.verticalConfigId,
         searcher: 'UNIVERSAL',
         entityId: this.associatedEntityId
