+++ static/js/HitchhikerJS.js
@@ -41,3 +41,6 @@ import AnswersExperience from './answers-experience';
 window.AnswersExperience = new AnswersExperience(runtimeConfig);
 
 export * from './video-apis';
+
+import DOM from './dom';
+export { DOM };
